US157414: List/list item: Improve performance on resize
