+++ components/list/test/list.visual-diff.html
@@ -624,6 +624,27 @@
 		</div>
 
 		<div style="width: 900px;">
+			<div class="visual-diff" id="breakpoint-list-item">
+				<d2l-list>
+					<d2l-list-item color="#00ff00" breakpoints="[1170, 391, 0, 0]">
+						<div style="background: blue; height: 400px; width: 400px;" slot="illustration"></div>
+						<d2l-list-item-content>
+							<div>Introductory Pirate Ipsum</div>
+							<div slot="supporting-info">Case shot Shiver me timbers gangplank crack Jennys tea cup ballast Blimey lee snow crow's nest rutters.</div>
+						</d2l-list-item-content>
+					</d2l-list-item>
+					<d2l-list-item>
+						<div style="background: blue; height: 400px; width: 400px;" slot="illustration"></div>
+						<d2l-list-item-content>
+							<div>Introductory Pirate Ipsum</div>
+							<div slot="supporting-info">Case shot Shiver me timbers gangplank crack Jennys tea cup ballast Blimey lee snow crow's nest rutters.</div>
+						</d2l-list-item-content>
+					</d2l-list-item>
+				</d2l-list>
+			</div>
+		</div>
+
+		<div style="width: 900px;">
 			<div class="visual-diff" id="breakpoint-842">
 				<d2l-list>
 					<d2l-list-item color="#00ff00">
