+++ components/list/test/list.visual-diff.js
@@ -229,7 +229,8 @@ describe('d2l-list', () => {
 			{ name: '842', selector: '#breakpoint-842' },
 			{ name: '636', selector: '#breakpoint-636' },
 			{ name: '580', selector: '#breakpoint-580' },
-			{ name: '0', selector: '#breakpoint-0' }
+			{ name: '0', selector: '#breakpoint-0' },
+			{ name: 'list item', selector: '#breakpoint-list-item' }
 		] },
 		{ category: 'dropdown', tests: [
 			{ name: 'open down', selector: '#dropdown-tooltips', action: () => openDropdown('#open-down'), after: () => closeDropdown('#open-down') }
